+++ packages/components/modules/messages/MessagesList/types.ts
@@ -3,7 +3,7 @@ import { FC } from 'react'
 import { VirtuosoProps } from 'react-virtuoso'
 
 import { MessagesListFragment$key } from '../../../__generated__/MessagesListFragment.graphql'
-import { MessagesGroupProps } from '../MessagesGroup/types'
+import { MessagesGroupProps } from './MessagesGroup/types'
 
 export interface MessagesListProps {
   roomRef: MessagesListFragment$key