+++ packages/components/modules/messages/graphql/queries/ChatRoomsQuery.ts
@@ -1,0 +1,14 @@
+import { graphql } from 'react-relay'
+
+export const ChatRoomsQuery = graphql`
+  query ChatRoomsQuery {
+    ...AllProfilesListFragment
+    me {
+      id
+      profile {
+        id
+        ...RoomsListFragment
+      }
+    }
+  }
+`