+++ packages/components/modules/messages/graphql/queries/MessageItem.ts
@@ -9,6 +9,7 @@ export const MessageItemFragment = graphql`
     inReplyTo {
       id
     }
+    isRead
     pk
     profile {
       id