+++ packages/components/modules/messages/index.ts
@@ -1,18 +1,29 @@
-export { default as MessageItem } from './MessageItem'
-export type * from './MessageItem/types'
-
 export { default as MessagesList } from './MessagesList'
 export type * from './MessagesList/types'
 
-export { default as MessageRoom } from './MessageRoom'
-export type * from './MessageRoom/types'
+export { default as ChatRoom } from './ChatRoom'
+export type * from './ChatRoom/types'
+
+export { default as ChatRoomsList } from './ChatRoomsList'
+export type * from './ChatRoomsList/types'
 
 export { default as SendMessage } from './SendMessage'
 export type * from './SendMessage/types'
 
+export { default as CreateChatRoomList } from './CreateChatRoomList'
+export type * from './CreateChatRoomList/types'
+
+export * from './context'
+
 export * from './graphql/mutations/SendMessage'
+export * from './graphql/mutations/CreateChatRoom'
+export * from './graphql/mutations/ReadMessages'
+
 export * from './graphql/queries/MessageItem'
 export * from './graphql/queries/MessagesList'
-export * from './graphql/queries/MessageRoomQuery'
+export * from './graphql/queries/ChatRoomQuery'
+export * from './graphql/queries/Room'
+export * from './graphql/queries/RoomsList'
 
 export { default as useMessagesListSubscription } from './graphql/subscriptions/useMessagesListSubscription'
+export { default as useMessageCountUpdateSubscription } from './graphql/subscriptions/useMessageCountUpdateSubscription'