+++ packages/components/modules/navigations/index.ts
@@ -22,4 +22,8 @@ export type * from './Header/AccountMenu/types'
 export { default as AccountPopover } from './Header/AccountMenu/AccountPopover'
 export type * from './Header/AccountMenu/AccountPopover/types'
 
+export { default as ViewportHeightContainer } from './ViewportHeightContainer'
+export type * from './ViewportHeightContainer/types'
+
+export * from './constants'
 export type * from './types'