+++ packages/wagtail/modules/components/Blocks/BannerBlock/BannerDescriptionRichText/index.tsx
@@ -5,7 +5,9 @@ import { RichTextBlockWrapper } from './styled'
 const BannerDescriptionRichText = ({ value }: { value: string }) => (
   <RichTextBlockWrapper>
     {/* TODO (Tech Debt): Sanitize the HTML content with, for example, DOMPurify. */}
-    <div dangerouslySetInnerHTML={{ __html: value }} /> {/* eslint-disable-line react/no-danger */}
+    <div
+      dangerouslySetInnerHTML={{ __html: value }}
+    /> {/* eslint-disable-line react/no-danger */}
   </RichTextBlockWrapper>
 )